Migrate shell sort
- Move implementation to header file
- Add missing curly braces
- Add to unit test file
- Update README
- Delete old shell_sort.cpp
