+++ C++/README.md
@@ -38,7 +38,7 @@ Implementation of well-known (and some rare) algorithms, in C++.
   * [Quick sort](include/algorithm/sorting/quick_sort.hpp) :white_check_mark:
   * [Radix sort](include/algorithm/sorting/radix_sort.hpp) 
   * [Selection sort](include/algorithm/sorting/selection_sort.hpp) :white_check_mark:
-  * [Shell sort](source/sorting/shell_sort.cpp)
+  * [Shell sort](include/algorithm/sorting/shell_sort.hpp) :white_check_mark:
 * String
   * [Longest common subsequence](include/algorithm/string/longest_common_subsequence.hpp) :white_check_mark:
   * Searching (pattern matching)
