+++ C++/test/algorithm/sorting/sorting.cpp
@@ -8,6 +8,7 @@
 #include "algorithm/sorting/merge_sort.hpp"
 #include "algorithm/sorting/quick_sort.hpp"
 #include "algorithm/sorting/selection_sort.hpp"
+#include "algorithm/sorting/shell_sort.hpp"
 
 // Prototypes
 int generate_random_int(int, int);
@@ -28,7 +29,8 @@ TEST_CASE("Sort in ascending order", "[sorting]") {
 			insertion_sort,
 			merge_sort,
 			quick_sort,
-			selection_sort
+			selection_sort,
+			shell_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
@@ -56,7 +58,8 @@ TEST_CASE("Sort in descending order", "[sorting]") {
 			insertion_sort,
 			merge_sort,
 			quick_sort,
-			selection_sort
+			selection_sort,
+			shell_sort
 	};
 
 	vector<int> original, algo_sorted, std_sorted;
